Refactor tests: rename packages
Rename packages in participants module for better organization
of supporting classes
