+++ rekolekcje-engine/src/verify/java/pl/oaza/warszawa/dor/rekolekcje/api/participants/storage/ParticipantsStorageBehaviour.java
@@ -1,4 +1,4 @@
-package pl.oaza.warszawa.dor.rekolekcje.api.participants.utils;
+package pl.oaza.warszawa.dor.rekolekcje.api.participants.storage;
 
 import pl.oaza.warszawa.dor.rekolekcje.api.core.Database;
 import pl.oaza.warszawa.dor.rekolekcje.api.participants.dto.ParticipantDTO;
