+++ rekolekcje-engine/src/verify/java/pl/oaza/warszawa/dor/rekolekcje/api/participants/utils/ParticipantsApiExpectations.java
@@ -5,6 +5,7 @@ import com.jayway.jsonpath.DocumentContext;
 import com.jayway.jsonpath.JsonPath;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.ResultActions;
+import pl.oaza.warszawa.dor.rekolekcje.api.participants.storage.ParticipantData;
 import pl.oaza.warszawa.dor.rekolekcje.api.participants.dto.ParticipantDTO;
 
 import java.io.UnsupportedEncodingException;
